refactor(#421): extract EditSheet+ExtraHeadersField from ProvidersPage (#466)
* refactor(#421): extract EditSheet+ExtraHeadersField to EditProviderSheet.tsx

- Create features/providers/components/EditProviderSheet.tsx (525 lines)
- ProvidersPage.tsx: 1033 → 518 lines (-515 lines)
- EditSheet (481 lines) + ExtraHeadersField (26 lines) now in own file

Closes #421

* fix(#421): deduplicate provider constants — import from lib/providers

- Create lib/providers.ts with PROVIDER_DISPLAY_NAMES + PROVIDER_SUGGESTED_MODELS
- EditProviderSheet + ProvidersPage both import the shared module
- Remove ~42 lines of duplicated constants

* chore: trigger CI re-run

---------
